requirements.txt
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
python-dotenv>=0.19.0
tqdm>=4.65.0
openai>=1.0.0
pandas>=1.5.0
scikit-learn>=1.0.0
spacy>=3.0.0 